requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
sqlalchemy==2.0.45
loguru==0.7.3
python-multipart==0.0.21
pytest==9.0.2
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.1